Starts with proper main func with logging; Remote debugging supported
Cleaned up initial boilerplate code from testing GPIO. Remote debugging
working now, required sudo access as well. Keyboard library runs,
but due to it running as root does not push the keypresses to the
ssh session but to local. Not an issue in the end as the app will be
run locally, but makes debugging trickier as navigation visually will
not match. When running through ssh you can manually hit the keys
which do register properly.