Add method for generating menus and include defaults
To keep lightweight and to stay entirely within one service (fim)
the menus are generated as individual images with the "selected"
option be the bolded on. Allows for the same code that traverses
through the images to be the same for the menus. Chose to
programmatically generate them as the Pillow library was already
needed for album art and then allows for easy updating in case of
future changes. Also may be used to generate and show track listing
while scrobbling is occuring.